feat: 리뷰 로그 관리 기능 추가 및 기존 코드 리팩토링 - 리뷰 로그 관리 클래스를 새로 추가하여 로그 ID 생성 및 로그 저장 기능을 통합함 - 기존의 save_review_log 함수를 ReviewLogManager로 대체하여 코드의 일관성을 높임 - 리뷰 로그 저장 시 비용 정보 추가 기능을 구현하여 로그의 유용성을 향상시킴 - 관련 테스트 케이스를 업데이트하여 새로운 구조에 맞춤
Refs: CR-18
