+++ tests/test_config_review_log_dir.py
@@ -293,7 +293,7 @@ class TestCLIReviewLogDirOption(unittest.TestCase):
     @patch("selvage.cli.get_model_info")
     def test_save_review_log_with_custom_dir(self, mock_get_model_info) -> None:
         """save_review_log 함수가 사용자 지정 디렉토리를 올바르게 처리하는지 테스트."""
-        from selvage.cli import save_review_log
+        from selvage.src.utils.logging.review_log_manager import ReviewLogManager
         from selvage.src.diff_parser.models.diff_result import DiffResult
         from selvage.src.models.model_provider import ModelProvider
         from selvage.src.models.review_status import ReviewStatus
@@ -316,12 +316,13 @@ class TestCLIReviewLogDirOption(unittest.TestCase):
         )
 
         # save_review_log 호출
-        log_path = save_review_log(
+        log_path = ReviewLogManager.save(
             prompt=None,
             review_request=review_request,
             review_response=None,
             status=ReviewStatus.SUCCESS,
             review_log_dir=self.custom_log_dir,
+            estimated_cost=None,
         )
 
         # 사용자 지정 디렉토리에 로그가 저장되었는지 확인
@@ -338,7 +339,7 @@ class TestCLIReviewLogDirOption(unittest.TestCase):
     @patch("selvage.cli.get_model_info")
     def test_save_review_log_with_default_dir(self, mock_get_model_info) -> None:
         """save_review_log 함수가 기본 디렉토리를 올바르게 사용하는지 테스트."""
-        from selvage.cli import save_review_log
+        from selvage.src.utils.logging.review_log_manager import ReviewLogManager
         from selvage.src.config import get_default_review_log_dir
         from selvage.src.diff_parser.models.diff_result import DiffResult
         from selvage.src.models.model_provider import ModelProvider
@@ -362,12 +363,13 @@ class TestCLIReviewLogDirOption(unittest.TestCase):
         )
 
         # save_review_log 호출 (review_log_dir=None)
-        log_path = save_review_log(
+        log_path = ReviewLogManager.save(
             prompt=None,
             review_request=review_request,
             review_response=None,
             status=ReviewStatus.SUCCESS,
             review_log_dir=None,
+            estimated_cost=None,
         )
 
         # 기본 디렉토리에 로그가 저장되었는지 확인
